[MODIFY] Main 화면 구성
