+++ src/app.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+
+import { Main } from './pages/Main';
 import { Todo } from './pages/todo/Todo';
 
 class App extends React.Component {
@@ -7,6 +9,7 @@ class App extends React.Component {
     return (
       <BrowserRouter>
         <Routes>
+          <Route exact path="/" element={<Main />} />
           <Route exact path="/todo" element={<Todo />} />
         </Routes>
       </BrowserRouter>
